# mathapp.apiDocs.yaml
openapi: '3.0.0'
info:
  version: "0.1.0"
  title: 'MathApp.API'
  description: 'This is the backend API for the ASU STEM Academy Math APP Project'
servers:
  - url: /api/v1
paths:
...
  /results:
    get:
      description: Get all results for all tests
      operationId: results.get_all
      responses:
        '200':
          description: All results
          content:
            application/json:
              schema:
                properties:
                  results:
                    type: array
                    items:
                      $ref: '#/components/schemas/result'
        '400':
          $ref: '#/components/responses/400Error'
        '404':
          $ref: '#/components/responses/404Error'
    post:
      description: Submit test results
      operationId: results.add_result
      requestBody:
        description: Submitted test results
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/new_result'
      responses:
        '201':
          description: Created
          headers:
            Location:
              schema:
                type: string
              description: Record locator
        '400':
          $ref: '#/components/responses/400Error'
  /results/{result_id}:
    get:
      description: Get result by result tracking ID
      operationId: results.get_by_result_id
      parameters:
        - name: result_id
          description: Result tracking ID number to look up
          in: path
          schema:
            type: integer
            minimum: 1
          required: true
      responses:
        '200':
          description: Aggregate Results
          content:
            application/json:
              schema:
                properties:
                  results:
                    type: object
                    $ref: '#/components/schemas/result'
        '400':
          $ref: '#/components/responses/400Error'
        '404':
          $ref: '#/components/responses/404Error'
  /results/{result_id}/summary:
    get:
      description: Get result summary by result tracking ID
      operationId: results.get_summary_by_result_id
      parameters:
        - name: result_id
          description: Result tracking ID number to look up
          in: path
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: Aggregate Results summary
          content:
            application/json:
              schema:
                properties:
                  results:
                    type: object
                    $ref: '#/components/schemas/result_summary'
        '400':
          $ref: '#/components/responses/400Error'
        '404':
          $ref: '#/components/responses/404Error'
...
  /results/student/{student_id}/summary:
    get:
      description: Get summary results for specified student
      operationId: results.get_summary_by_student
      parameters:
        - name: student_id
          description: Student ID to lookup
          in: path
          schema:
            type: integer
            minimum: 1
          required: true
      responses:
        '200':
          description: Aggregate Results
          content:
            application/json:
              schema:
                properties:
                  results:
                    type: array
                    items:
                      $ref: '#components/schemas/result_summary'
        '400':
          $ref: '#/components/responses/400Error'
        '404':
          $ref: '#/components/responses/404Error'
  /results/teacher/{teacher_id}:
    get:
      description: Get results for all tests by the specified teacher
      operationId: results.get_by_teacher
      parameters:
        - name: teacher_id
          description: Teacher ID to lookup
          in: path
          schema:
            type: integer
            minimum: 1
          required: true
      responses:
        '200':
          description: Aggregate Results
          content:
            application/json:
              schema:
                properties:
                  results:
                    type: array
                    items:
                      $ref: '#/components/schemas/result'
        '400':
          $ref: '#/components/responses/400Error'
        '404':
          $ref: '#/components/responses/404Error'
  /results/teacher/{teacher_id}/summary:
    get:
      description: Get summary results for specified teacher
      operationId: results.get_summary_by_teacher
      parameters:
        - name: teacher_id
          description: Teacher ID to lookup
          in: path
          schema:
            type: integer
            minimum: 1
          required: true
      responses:
        '200':
          description: Aggregate Results
          content:
            application/json:
              schema:
                properties:
                  results:
                    type: array
                    items:
                      $ref: '#components/schemas/result_summary'
        '400':
          $ref: '#/components/responses/400Error'
        '404':
          $ref: '#/components/responses/404Error'
  /results/test/{test_id}:
    get:
      description: Get results for a specific test by the test ID
      operationId: results.get_by_test_id
      parameters:
        - name: test_id
          description: Test tracking ID number to look up the results for
          in: path
          schema:
            type: integer
            minimum: 1
          required: true
      responses:
        '200':
          description: Aggregate Results
          content:
            application/json:
              schema:
                properties:
                  results:
                    type: array
                    items:
                      $ref: '#/components/schemas/result'
        '400':
          $ref: '#/components/responses/400Error'
        '404':
          $ref: '#/components/responses/404Error'
  /results/test/{test_id}/summary:
    get:
      description: Get summary results for a specific test by the test ID
      operationId: results.get_summary_by_test_id
      parameters:
        - name: test_id
          description: Test tracking ID number to look up the results for
          in: path
          schema:
            type: integer
            minimum: 1
          required: true
      responses:
        '200':
          description: Aggregate results
          content:
            application/json:
              schema:
                properties:
                  results:
                    type: array
                    items:
                      $ref: '#components/schemas/result_summary'
        '400':
          $ref: '#/components/responses/400Error'
        '404':
          $ref: '#/components/responses/404Error'
# Reusable Definitions     
components:
  schemas:
...
    result:
      type: object
      properties:
        result_id:
          type: integer
          description: Tracking number for test attempt submission
        time_taken:
          type: string
          description: Time taken for the individual test attempt ("HH:MM:SS")
        testId:
          type: integer
          description: Test ID of test this submission is for
        answers:
          type: array
          items:
            properties:
              student_answer:
                type: integer
                description: Answer student submitted 
              correctly_answered:
                type: boolean
                description: Whether the student's answer was correct
              operation:
                type: string
                description: +, -, *, /
              operand_1:
                type: integer
                description: Operand preceding the operation
              operand_2:
                type: integer
                description: Operand subequent to operation
    result_summary:
      type: object
      properties:
        result_id:
          type: integer
          description: Tracking number for test attempt submission
        time_taken:
          type: string
          description: Time taken for the individual test attempt ("HH:MM:SS")
        testId:
          type: integer
          description: Test ID of test this submission is for
        total_questions:
          type: integer
          description: Number of questions in the test attempt
        correctly_answered:
          type: integer
          description: Number of questions the student answered correctly
    new_result:
      type: object
      properties:
        time_taken:
          type: string
          description: Time taken for the individual test attempt ("HH:MM:SS")
        testId:
          type: integer
          description: Test ID of the test the result belongs to
        answers:
          type: array
          items:
            properties:
              student_answer:
                type: integer
                description: Answer the student submitted
              operation:
                type: string
                description: +, -, *, /
              operand_1:
                type: integer
                description: Operand preceding the operation
              operand_2:
                type: integer
                description: Operand subsequent to operation
    error:
      type: object
      properties:
        message:
          type: string
          description: Error Message from server
...
  responses:
    400Error:
      description: Invalid Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/error'
    404Error:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/error'